Módulo pretty para sintaxe do mysql Segregação da classe MySqlTranslator
